services/media-service/requirements.txt
-r ../../requirements-shared.txt
Pillow==10.0.0
cloudinary==1.34.0
qrcode[pil]==7.4.2
reportlab==4.0.4
gunicorn==21.2.0
